+++ runtime/calamari/src/lib.rs
@@ -50,7 +50,7 @@ use frame_support::{
 };
 use frame_system::{
     limits::{BlockLength, BlockWeights},
-    EnsureRoot, RawOrigin,
+    EnsureRoot,
 };
 use manta_primitives::{
     constants::{time::*, RocksDbWeight, STAKING_PALLET_ID, TREASURY_PALLET_ID, WEIGHT_PER_SECOND},
@@ -130,10 +130,10 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("calamari"),
     impl_name: create_runtime_str!("calamari"),
     authoring_version: 2,
-    spec_version: 4020,
+    spec_version: 4030,
     impl_version: 1,
     apis: RUNTIME_API_VERSIONS,
-    transaction_version: 10,
+    transaction_version: 11,
     state_version: 0,
 };
 
@@ -884,7 +884,7 @@ pub type CheckedExtrinsic = generic::CheckedExtrinsic<AccountId, Call, SignedExt
 
 /// Types for runtime upgrading.
 /// Each type should implement trait `OnRuntimeUpgrade`.
-pub type OnRuntimeUpgradeHooks = (migrations::asset_id::AssetIdMigration<Runtime>,);
+pub type OnRuntimeUpgradeHooks = ();
 /// Executive: handles dispatch to the various modules.
 pub type Executive = frame_executive::Executive<
     Runtime,
@@ -1282,24 +1282,4 @@ cumulus_pallet_parachain_system::register_validate_block! {
     Runtime = Runtime,
     BlockExecutor = pallet_author_inherent::BlockExecutor::<Runtime, Executive>,
     CheckInherents = CheckInherents,
-}
-
-pub struct MantaPaySuspensionManager;
-impl pallet_manta_pay::SuspendMantaPay for MantaPaySuspensionManager {
-    fn suspend_manta_pay_execution() {
-        match TransactionPause::pause_transactions(
-            RawOrigin::Root.into(),
-            vec![(
-                b"MantaPay".to_vec(),
-                vec![
-                    b"to_private".to_vec(),
-                    b"private_transfer".to_vec(),
-                    b"to_public".to_vec(),
-                ],
-            )],
-        ) {
-            Ok(_) => log::error!("MantaPay has been suspended due to an unexpected internal ledger error!"),
-            Err(tx_pause_error) => log::error!("MantaPay encountered an unexpected internal ledger error, but failed to be suspended with the following tx-pause error: {:?}!", tx_pause_error)
-        }
-    }
 }